DEV-189 #time 1h S3 using S3R SD write functions and sensor handling

+++ LogAndStream/Shimmer_Driver/ICM20948/ICM20948.h
@@ -196,8 +196,8 @@ uint8_t ICM20948_isMagDataRdy(void);
 void ICM20948_getMag(uint8_t *buf);
 
 uint8_t ICM20948_isMagSampleSkipEnabled(void);
-uint8_t ICM20948_hasTimeoutPeriodPassed(uint32_t currentSampleTsTicks);
-uint8_t ICM20948_getMagAndStatus(uint32_t currentSampleTsTicks, uint8_t *buf);
+uint8_t ICM20948_hasTimeoutPeriodPassed(uint64_t currentSampleTsTicks);
+uint8_t ICM20948_getMagAndStatus(uint64_t currentSampleTsTicks, uint8_t *buf);
 
 //read the x, y and z mag sensitivity adjustment values
 void ICM20948_getMagSensitivityAdj(uint8_t *buf);
